moved Token Generation Test to JWT Test
