+++ src/main/java/com/metamug/mason/dao/AuthDAO.java
@@ -525,9 +525,8 @@ import org.json.JSONObject;
  */
 public class AuthDAO {
 
-    DataSource ds;
-    private static final String STATUS = "status";
-    private static final int EXPIRY_DAYS = 90;
+    private DataSource ds;
+    public static final String STATUS = "status";
 
     public AuthDAO(DataSource ds) {
         this.ds = ds;
@@ -559,41 +558,6 @@ public class AuthDAO {
             Logger.getLogger(AuthDAO.class.getName()).log(Level.SEVERE, ex.getMessage(), ex);
         }
         return status;
-    }
-
-    /**
-     * This is used by to generate the token.
-     *
-     * @param user username to be matched in the auth query
-     * @param pass password to be matched in auth query
-     * @param authQuery The Bearer Auth query in web.xml
-     * @return
-     */
-    public JSONObject getBearerDetails(String user, String pass, String authQuery) {
-        JSONObject jwtPayload = new JSONObject();
-        jwtPayload.put(STATUS, 0);
-        try (Connection con = ds.getConnection()) {
-            if (!authQuery.isEmpty()) {
-                try (PreparedStatement stmt = con.prepareStatement(authQuery.replaceAll("\\$(\\w+(\\.\\w+){0,})", "? "))) {
-                    stmt.setString(1, user);
-                    stmt.setString(2, pass);
-                    try (ResultSet result = stmt.executeQuery()) {
-                        JSONArray audArray = new JSONArray();
-                        while (result.next()) {
-                            jwtPayload.put("sub", result.getString(1));
-                            audArray.put(result.getString(2));
-                        }
-                        jwtPayload.put("aud", audArray);
-                        LocalDateTime ldt = LocalDateTime.now().plusDays(EXPIRY_DAYS);
-//                        LocalDateTime.now();
-                        jwtPayload.put("exp", ldt.toEpochSecond(ZoneOffset.UTC)); //this needs to be configured
-                    }
-                }
-            }
-        } catch (SQLException ex) {
-            Logger.getLogger(AuthDAO.class.getName()).log(Level.SEVERE, ex.getMessage(), ex);
-        }
-        return jwtPayload;
     }
 
 }
